+++ psyrun/backend/base.py
@@ -4,6 +4,66 @@ import os
 import sys
 
 
+class JobSourceFile(object):
+    """Describes a source code file for a job.
+
+    Parameters
+    ----------
+    path : str
+        Path to write the source file to.
+    task : Task
+        Task the source file corresponds to.
+    job_code : str
+        Job specific code to execute.
+
+    Attributes
+    ----------
+    written : bool
+        Whether the source file has been written.
+    full_code : str
+        The full job code including non job-specific parts that are used for
+        every job.
+    """
+
+    def __init__(self, path, task, job_code):
+        self.path = path
+        self.task = task
+        self.job_code = job_code
+        self.written = False
+
+    def write(self):
+        """Write the job code to the file *self.path*."""
+        with open(self.path, 'w') as f:
+            f.write(self.full_code)
+        self.written = True
+
+    @property
+    def full_code(self):
+        return '''
+try:
+    import faulthandler
+    faulthandler.enable()
+except:
+    pass
+
+import time
+print("")
+print("----------------------------------------------------------------------")
+print("Job started ({{}})".format(time.strftime('%a, %d %b %Y %H:%M:%S, %Z')))
+print("----------------------------------------------------------------------")
+
+import os
+os.chdir({taskdir!r})
+
+from psyrun.tasks import TaskDef
+task = TaskDef({taskpath!r})
+{code}
+        '''.format(
+            path=sys.path,
+            taskdir=os.path.abspath(os.path.dirname(self.task.path)),
+            taskpath=os.path.abspath(self.task.path), code=self.job_code)
+
+
 class Backend(object):
     """Abstract base class for processing backends.
 
@@ -31,51 +91,56 @@ class Backend(object):
         if not os.path.exists(self.workdir):
             os.makedirs(self.workdir)
 
-    def submit(self, code, name, depends_on=None):
+    def submit_code(self, code, name, depends_on=None, args=None):
         """Submits some code to execute to the task scheduler.
 
         Parameters
         ----------
-        code: str
+        code : str
             Code to execute in job.
-        name: str
+        name : str
             Job name.
-        depends_on: sequence
+        depends_on : sequence
             Job IDs that have to finish before the submitted code can be
             executed.
+        args : sequence
+            Additional arguments to pass to the job.
 
         Returns
         -------
         dict
             Contains the id of the submitted job under the key ``'id'``.
         """
-        code = '''
-try:
-    import faulthandler
-    faulthandler.enable()
-except:
-    pass
-
-import time
-print("")
-print("----------------------------------------------------------------------")
-print("Job started ({{}})".format(time.strftime('%a, %d %b %Y %H:%M:%S, %Z')))
-print("----------------------------------------------------------------------")
-
-import os
-os.chdir({taskdir!r})
-
-from psyrun.tasks import TaskDef
-task = TaskDef({taskpath!r})
-{code}
-        '''.format(
-            path=sys.path,
-            taskdir=os.path.abspath(os.path.dirname(self.task.path)),
-            taskpath=os.path.abspath(self.task.path), code=code)
         codefile = os.path.join(self.workdir, name + '.py')
+        return self.submit_file(JobSourceFile(codefile, self.task, code),
+                                name, depends_on=depends_on, args=args)
+
+    def submit_file(self, job_source_file, name, depends_on=None, args=None):
+        """Submits a source file to execute to the task scheduler.
+
+        Parameters
+        ----------
+        job_source_file : `JobSourceFile`
+            Source file to execute in job.
+        name: str
+            Job name.
+        depends_on: sequence
+            Job IDs that have to finish before the submitted code can be
+            executed.
+        args : sequence
+            Additional arguments to pass to the job.
+
+        Returns
+        -------
+        dict
+            Contains the id of the submitted job under the key ``'id'``.
+        """
+        if args is None:
+            args = []
+
         output_filename = os.path.join(self.workdir, name + '.log')
-        with open(codefile, 'w') as f:
-            f.write(code)
+        if not job_source_file.written:
+            job_source_file.write()
 
         for job in self.task.scheduler.get_jobs():
             status = self.task.scheduler.get_status(job)
@@ -83,8 +148,8 @@ task = TaskDef({taskpath!r})
                 self.task.scheduler.kill(job)
 
         return self.task.scheduler.submit(
-            [self.task.python, codefile], output_filename, name, depends_on,
-            self.task.scheduler_args)
+            [self.task.python, job_source_file.path] + args, output_filename,
+            name, depends_on, self.task.scheduler_args)
 
     def create_job(self, cont=False):
         """Create the job tree to process given task.
@@ -105,3 +170,17 @@ task = TaskDef({taskpath!r})
         the task definition, but that have not been evaluated yet.
         """
         raise NotImplementedError
+
+    def get_queued(self):
+        """Returns parameter sets that are still in queue to be processed.
+
+        May return ``None`` if this is not supported by the backend.
+        """
+        raise NotImplementedError()
+
+    def get_failed(self):
+        """Returns a list of failed jobs.
+
+        May return ``None`` if this is not supported by the backend.
+        """
+        raise NotImplementedError()